installed github-api. testing receiving of repos and issues.
just testing for now. We'll want to remove this out of the component layer.
The good news, is we were able to successfully pull down a list of repos
and issues with just the api key. So from here on out, we'll be mostly
working on scaffolding mobx out, adding the local storage / session storage,
and lastly addressing UI